refactor: refactor configs
