+++ app/(frontend)/(with-sidebar)/flights/search/page.client.tsx
@@ -15,7 +15,7 @@ import {
   SearchForm,
   type SearchFormData,
 } from "@/components/flights/search";
-import { SEAT_CLASS_TYPE_MAP } from "@/config/ui/flights";
+import { SEAT_CLASS_TYPE_MAP } from "@/config/ui";
 import { useFlightSearchState } from "@/hooks/use-flight-search-state";
 import { buildFlightSearchUrl } from "@/lib/flights/search-params";
 import { dateToLocalDateString } from "@/lib/format/date";
