+++ app/_actions/auth.ts
@@ -4,12 +4,12 @@ import { revalidatePath } from "next/cache";
 import { headers } from "next/headers";
 
 import { requireSessionUser } from "@/actions/session";
-import { auth } from "@/domains/auth";
 import {
   changePassword,
   setPasswordForOAuthUser,
   unlinkSocialAccount,
 } from "@/domains/auth/auth.service";
+import { auth } from "@/infra/auth";
 import { logger } from "@/infra/logging";
 import { validateAccount } from "@/lib/validation/account";
 import {
