+++ app/_actions/user.ts
@@ -3,13 +3,13 @@
 import { headers } from "next/headers";
 
 import { getSessionUser, requireSessionUser } from "@/actions/session";
-import { auth } from "@/domains/auth";
 import {
   getUserBalance,
   getUserInfo,
   getUserSecurityStatus,
 } from "@/domains/user/user.read.service";
 import { rechargeBalance, updateUserInfo } from "@/domains/user/user.service";
+import { auth } from "@/infra/auth";
 import type { ActionResult } from "@/types/common";
 import type { UserInfo, UserSecurityStatus } from "@/types/dto";
 import {
