+++ app/_components/flights/search/FlightSearchHeader.tsx
@@ -15,7 +15,7 @@ import {
   TooltipContent,
   TooltipTrigger,
 } from "@/components/ui/tooltip";
-import { FLIGHT_UI_TEXT, ROUND_TRIP_STEPS } from "@/config/ui/flights";
+import { FLIGHT_UI_TEXT, ROUND_TRIP_STEPS } from "@/config/ui";
 import { formatDateWithWeekday, formatTime } from "@/lib/format/date";
 import type { CityData } from "@/types/dto";
 import type { TripType } from "@/types/validations";
