+++ app/api/auth/[...all]/route.ts
@@ -1,5 +1,5 @@
 import { toNextJsHandler } from "better-auth/next-js";
 
-import { auth } from "@/domains/auth";
+import { auth } from "@/infra/auth";
 
 export const { POST, GET } = toNextJsHandler(auth);
