--- src/config/docs/heading.config.ts
@@ -1,6 +1,0 @@
-export {
-  colorfulHeadingStyles,
-  headingStyles,
-  minimalHeadingStyles,
-  sidebarHeadingStyles,
-} from "@/config/ui/heading";
