+++ src/config/docs/index.ts
@@ -1,7 +1,1 @@
-export { getLLMText } from "./get-llm-text";
-export {
-  ProvideAIAnnotationsToolSchema,
-  ProvideLinksToolSchema,
-} from "./inkeep-qa-schema";
-export { baseOptions } from "./layout.shared";
-export { getPageImage, source } from "./source";
+export * from "./layout-shared.config";
