+++ src/domains/auth/index.ts
@@ -7,5 +7,3 @@ const authClient = createAuthClient({
 });
 
 export const { signIn, signUp, useSession } = authClient;
-
-export { auth } from "@/infra/auth";
