--- src/domains/booking/booking.error.ts
@@ -1,1 +1,0 @@
-export { bookingErrorConfigs } from "@/config/errors/booking";
