--- src/domains/booking/index.ts
@@ -1,8 +1,0 @@
-export { getOrderForAncillary } from "./ancillary.repository";
-export {
-  getFlightSeatClassById,
-  getFlightSeatClassesByIds,
-  getSavedPassengers,
-} from "./booking-passengers.repository";
-export { getOrderConfirmation } from "./confirmation.repository";
-export { getOrderForPayment } from "./payment.repository";
