--- src/domains/flights/flights.error.ts
@@ -1,1 +1,0 @@
-export { flightErrorConfigs } from "@/config/errors/flight";
