--- src/domains/flights/flights.read.service.ts
@@ -1,3 +1,0 @@
-export { getAllCities } from "@/domains/flights/city.repository";
-export { searchFlightsWithHistory } from "@/domains/flights/flight-search.service";
-export { getRecentSearchHistory } from "@/domains/flights/flight-search-history.repository";
