+++ src/infra/fumadocs/source.ts
@@ -18,11 +18,3 @@ export function getPageImage(page: InferPageType<typeof source>) {
     url: `/og/docs/${segments.join("/")}`,
   };
 }
-
-export async function getLLMText(page: InferPageType<typeof source>) {
-  const processed = await page.data.getText("processed");
-
-  return `# ${page.data.title} (${page.url})
-
-${processed}`;
-}
