+++ src/lib/auth/helpers.ts
@@ -1,7 +1,7 @@
 import { headers } from "next/headers";
 import { redirect } from "next/navigation";
 
-import { auth } from "@/domains/auth";
+import { auth } from "@/infra/auth";
 
 /**
  * Result type for authentication operations
