--- src/lib/server/index.ts
@@ -1,3 +1,0 @@
-export * from "./uuid";
-export * from "@/infra/http";
-export { logger } from "@/infra/logging";
